Change which migration-generated files we track.
Added are the delayed_jobs and rails_settings created files, as they are
not likely to change until we pull in a newer version of the gem.

Gone is rails/db/schema.rb, because it is likely to grow more sections
as we start doing builds with workloads, and I do not think the
cognitive overhead for tracking when it shold be committed vs. when it
should be ignored is worth it.

This also consolidates the .gitignore files.