Make IP address allocation avoid deadlocks harder.
Make sure we get an exclusive lock on the network_allocations table
before actually trying to allocate an IP, and retry any failed
transactions that happen due to deadlocks.  This will keep the system
from trying to hand out the same IP in different transactions that
happen to be running in parallel.