Add our own transaction method.
It makes serializable the default isolation level for a transaction,
retries transactions automatically on serialization failures, and
tries up to 3 times to resolve deadlock-related failures.

It also allows us to declare that a transaction should be read-only,
where we want to read data in a fashion that is compatible with
everything else being serialized.

We also add a locked_transaction method that takes a table lock on the
table associated with the class.