Simplify enforce_acyclic when creating role_requires.
Instead of manually iterating over all the role_requires to find
conflicting ones, fix up the all_role_requires view to have the
functionality we need to do it in one shot.

If we find a cyclic role dependency, we will spit out all paths through
the role dependency graph that would be made cyclic.