Use golden Sledgehammer image.
Instead of making it be a requirement that everyone build their own
Sledgehammer locally, arrange for the bootstrap recipe to download the
appropriate Sledgehammer image from our bucket on Amazon, using sha1
hash checking along the way to ensure that everything is peachy.

Sledgehammer images are indexed by the sha1sum of their source files.
See the changes in tools/build_sledgehammer.sh for more details.